move router to etna-js
